+++ database/migrations/2019_03_22_052957_create_reports_table.php
@@ -22,7 +22,7 @@ class CreateReportsTable extends Migration
             $table->integer('user_id');
             $table->text('comment')->nullable();
             $table->integer('supervisor_id')->nullable();
-            $table->integer('total_average')->nullable();
+            $table->float('total_average')->nullable();
             $table->boolean('forAssessment')->nullable();
             $table->boolean('assessed')->nullable();
             $table->integer('oldreport_id')->nullable();
